# openapi.yaml
openapi: 3.0.3
info:
  title: Taskflow API
  description: >
    Durable job scheduling and queue service. PostgreSQL stores job definitions,
    execution history, and callback-delivery state while Redis/BullMQ provides
    delayed, recurring, retry, and completion-callback execution. Production API
    callers authenticate as named clients with least-privilege scopes.
  version: "1.4"
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
  /health:
    get:
      security: []
      summary: Lightweight health check
      responses:
        "200":
          description: Process is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveHealth"
  /health/live:
    get:
      security: []
      summary: Liveness probe
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveHealth"
  /health/ready:
    get:
      security: []
      summary: Dependency readiness probe
      description: Checks PostgreSQL and Redis with bounded timeouts.
      responses:
        "200":
          description: Runtime dependencies are reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessHealth"
        "503":
          description: One or more runtime dependencies are unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessHealth"
  /v1/auth/whoami:
    get:
      summary: Inspect the authenticated API client
      description: Returns the client id and configured scopes for the current bearer credential.
      responses:
        "200":
          description: Authenticated client identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs:
    post:
      summary: Create a job
      description: >
        Requires `jobs.write`. Creates a one-off or recurring job. If
        `idempotencyKey` is provided and a job with that key already exists, the
        existing durable job is returned and its BullMQ projection is repaired if
        necessary. An optional `callbackUrl` requests a signed asynchronous
        notification after each completed firing (success or final failed attempt).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            examples:
              delayedOnce:
                summary: One-off job
                value:
                  type: log_message
                  payload: { message: "run this later" }
                  schedule: { type: once, runAt: "2026-09-12T20:00:00Z" }
                  idempotencyKey: demo-once-1
                  maxAttempts: 3
              withCallback:
                summary: One-off job with completion callback
                value:
                  type: log_message
                  payload: { message: "notify me" }
                  schedule: { type: once }
                  callbackUrl: https://api.example.com/taskflow/events
              recurring:
                summary: Recurring job, every day at 2am UTC
                value:
                  type: log_message
                  payload: { message: "nightly cleanup tick" }
                  schedule: { type: recurring, cron: "0 2 * * *", timezone: "UTC" }
      callbacks:
        completion:
          '{$request.body#/callbackUrl}':
            post:
              security: []
              summary: Taskflow completion callback
              description: >
                Sent after a successful execution or the final failed attempt.
                Consumers should verify the HMAC signature against the exact raw
                request body and deduplicate by `x-taskflow-delivery-id`.
              parameters:
                - name: x-taskflow-delivery-id
                  in: header
                  required: true
                  schema: { type: string }
                - name: x-taskflow-signature
                  in: header
                  required: true
                  description: sha256=<HMAC-SHA256 hex digest of the raw body>
                  schema: { type: string }
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/CompletionCallbackEvent"
              responses:
                "200":
                  description: Callback accepted by the consumer
                "204":
                  description: Callback accepted with no response body
      responses:
        "201":
          description: Job created, or existing idempotent job returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: Validation error or callback destination not allowed
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      summary: List jobs
      description: Requires `jobs.read`.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/JobStatus"
        - name: type
          in: query
          schema: { type: string }
        - name: scheduleType
          in: query
          schema:
            type: string
            enum: [ONCE, RECURRING]
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Jobs ordered by creation time descending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: Invalid filter
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{id}:
    get:
      summary: Get a job and recent execution history
      description: Requires `jobs.read`. Includes the 20 most recent executions for convenience. Use the executions endpoint for full history.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Job not found
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{id}/executions:
    get:
      summary: List job execution history
      description: >
        Requires `jobs.read`. Returns execution attempts newest first using opaque
        cursor pagination. Supply the previous response's `pageInfo.nextCursor`
        to continue.
      parameters:
        - $ref: "#/components/parameters/JobId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ExecutionStatus"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page.
          schema: { type: string, maxLength: 1024 }
      responses:
        "200":
          description: Cursor-paginated execution history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Job not found
        "422":
          description: Invalid status, limit, or cursor
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{id}/cancel:
    post:
      summary: Cancel a scheduled job
      description: >
        Requires `jobs.write`. Atomically marks a SCHEDULED job CANCELLED in
        PostgreSQL before best-effort BullMQ cleanup. RUNNING and terminal jobs
        return a conflict rather than reporting a cancellation that cannot be
        guaranteed.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: The cancelled job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Job not found
        "409":
          description: Job is running or already terminal
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/operations/overview:
    get:
      summary: Runtime operations overview
      description: Requires `operations.read`. Returns job/callback BullMQ counts, configured handler limits, durable status counts, and process uptime.
      responses:
        "200":
          description: Current runtime overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationsOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Production tokens use `<clientId>.<secret>` from TASKFLOW_API_CLIENTS.
        Client scopes are enforced by the API after authentication.
  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid bearer credential
    Forbidden:
      description: Authenticated client lacks the required scope
    RateLimited:
      description: API rate limit exceeded
  schemas:
    LiveHealth:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    ReadinessHealth:
      type: object
      required: [status, checks]
      properties:
        status: { type: string, enum: [ready, not_ready] }
        checks:
          type: object
          required: [postgres, redis]
          properties:
            postgres: { type: string, enum: [ok, failed] }
            redis: { type: string, enum: [ok, failed] }
    AuthIdentity:
      type: object
      required: [clientId, scopes]
      properties:
        clientId: { type: string }
        scopes:
          type: array
          items:
            type: string
            enum: [jobs.read, jobs.write, operations.read, "*"]
    JobStatus:
      type: string
      enum: [SCHEDULED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    ExecutionStatus:
      type: string
      enum: [RUNNING, SUCCEEDED, FAILED]
    ScheduleOnce:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [once] }
        runAt:
          type: string
          format: date-time
          description: Omit to run as soon as possible.
    ScheduleRecurring:
      type: object
      required: [type, cron]
      properties:
        type: { type: string, enum: [recurring] }
        cron: { type: string, example: "0 2 * * *" }
        timezone: { type: string, default: UTC, example: "America/Chicago" }
    CreateJobRequest:
      type: object
      required: [type, payload, schedule]
      properties:
        type:
          type: string
          description: Must match a registered handler such as log_message, http_request, or simulate_failure.
        payload:
          description: Handler-specific JSON input validated by the handler.
        schedule:
          oneOf:
            - $ref: "#/components/schemas/ScheduleOnce"
            - $ref: "#/components/schemas/ScheduleRecurring"
        priority: { type: integer, minimum: 0, maximum: 10, default: 0 }
        maxAttempts: { type: integer, minimum: 1, maximum: 20, default: 5 }
        idempotencyKey: { type: string, minLength: 1, maxLength: 200 }
        callbackUrl:
          type: string
          format: uri
          maxLength: 2048
          description: Optional signed POST destination for final execution outcomes.
    Job:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        payload: {}
        scheduleType: { type: string, enum: [ONCE, RECURRING] }
        runAt: { type: string, format: date-time, nullable: true }
        cronExpression: { type: string, nullable: true }
        timezone: { type: string }
        status:
          $ref: "#/components/schemas/JobStatus"
        priority: { type: integer }
        maxAttempts: { type: integer }
        attemptCount: { type: integer }
        idempotencyKey: { type: string, nullable: true }
        callbackUrl: { type: string, format: uri, nullable: true }
        lastError: { type: string, nullable: true }
        lastRunAt: { type: string, format: date-time, nullable: true }
        nextRunAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobExecution:
      type: object
      properties:
        id: { type: string }
        attemptNumber: { type: integer }
        status:
          $ref: "#/components/schemas/ExecutionStatus"
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
        durationMs: { type: integer, nullable: true }
        result: { nullable: true }
        error: { type: string, nullable: true }
    CompletionCallbackEvent:
      type: object
      required: [event, deliveryId, sentAt, job, execution]
      properties:
        event:
          type: string
          enum: [job.execution.succeeded, job.execution.failed]
        deliveryId: { type: string }
        sentAt: { type: string, format: date-time }
        job:
          type: object
          required: [id, type, scheduleType, status]
          properties:
            id: { type: string }
            type: { type: string }
            scheduleType: { type: string, enum: [ONCE, RECURRING] }
            status:
              $ref: "#/components/schemas/JobStatus"
        execution:
          type: object
          required: [id, attemptNumber, status, startedAt]
          properties:
            id: { type: string }
            attemptNumber: { type: integer }
            status:
              $ref: "#/components/schemas/ExecutionStatus"
            startedAt: { type: string, format: date-time }
            finishedAt: { type: string, format: date-time, nullable: true }
            durationMs: { type: integer, nullable: true }
            result: { nullable: true }
            error: { type: string, nullable: true }
    ExecutionPage:
      type: object
      required: [data, pageInfo]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/JobExecution"
        pageInfo:
          type: object
          required: [nextCursor, hasMore]
          properties:
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            executions:
              type: array
              maxItems: 20
              items:
                $ref: "#/components/schemas/JobExecution"
    OperationsOverview:
      type: object
      required: [timestamp, process, queue, callbacks, handlerLimits, jobs]
      properties:
        timestamp: { type: string, format: date-time }
        process:
          type: object
          properties:
            uptimeSeconds: { type: integer, minimum: 0 }
        queue:
          type: object
          description: Main job queue counts by BullMQ state.
          additionalProperties:
            type: integer
        callbacks:
          type: object
          required: [queue, deliveries]
          properties:
            queue:
              type: object
              additionalProperties:
                type: integer
            deliveries:
              type: object
              description: Durable CallbackDelivery counts by PENDING, DELIVERED, and FAILED status.
              additionalProperties:
                type: integer
        handlerLimits:
          type: object
          required: [concurrency, rate]
          properties:
            concurrency:
              type: object
              additionalProperties:
                type: integer
            rate:
              type: object
              additionalProperties:
                type: object
                required: [max, windowMs]
                properties:
                  max: { type: integer }
                  windowMs: { type: integer }
        jobs:
          type: object
          additionalProperties:
            type: integer